and one more "int" which should be "float" ( at the moment )

+++ source/glest_game/graphics/particle_type.h
@@ -66,7 +66,7 @@ protected:
 	float sizeNoEnergy;
 	float speed;
 	float gravity;
-	int emissionRate;
+	float emissionRate;
 	int energyMax;
 	int energyVar;
 	string mode;
